Fix duplicated files
