--- solutions/devsprint-leonardo-santos-1/FinanceAppTests/Modules/Home/Doubles/FinanceServiceSpy.swift
@@ -1,14 +1,0 @@
-@testable import FinanceApp
-
-final class FinanceServiceHomeDataSpy: FinanceServiceHomeDataProtocol {
-    
-    private(set) var fetchHomeDataTriggered: Bool = false
-    private(set) var fecthHomeDataCount = 0
-    var fetchHomeDataReturned: HomeData?
-    
-    func fetchHomeData(_ completion: @escaping (HomeData?) -> Void) {
-        fetchHomeDataTriggered = true
-        fecthHomeDataCount += 1
-        completion(fetchHomeDataReturned)
-    }
-}
